+++ src/pdf/GiftCard/App.js
@@ -2,17 +2,7 @@ import React from 'react';
 import Component from './Component';
 
 import { formatDynamicAmount } from '../../lib/formatCurrency'
-
-const fullMonths = [
-  'January', 'February', 'March', 'April', 'May', 'June', 'July', 'August', 'September', 'October', 'November', 'December',
-]
-
-const formatDate = (date) => {
-  const monthIdx = date.getMonth()
-  const monthStr = fullMonths[monthIdx]
-
-  return `${date.getDate()} ${monthStr} ${date.getFullYear()}`
-}
+import { formatDate } from '../../lib/formatDate'
 
 export default (props) => {
   const giftCardAmount = formatDynamicAmount(props.gift_card_value, props.currency)
